Add tokenizer; Move away from Nom (#2)
Parsing LocalCache.java has dropped from 50ms to 10ms.

Mainly because we don't process comments by characters. This is especially significant when we try to parse different alternatives (e.g. try parsing method, try parsing field in class body). We don't want to repeatedly process the prefix comments.

The tokenizing stage categorizes comments and discard them (because we don't need them at the moment). When we don't discard the comments, it only adds a few millis to the time.